exceptionHandling la answer requests

+++ src/main/java/com/ibm/ro/tm/apprenticeship/poll/metter/service/AnswerService.java
@@ -10,11 +10,14 @@ import org.springframework.stereotype.Service;
 import com.ibm.ro.tm.apprenticeship.poll.metter.entity.Answer;
 import com.ibm.ro.tm.apprenticeship.poll.metter.entity.Poll;
 import com.ibm.ro.tm.apprenticeship.poll.metter.entity.User;
+import com.ibm.ro.tm.apprenticeship.poll.metter.exception.AnswerNotFoundException;
 import com.ibm.ro.tm.apprenticeship.poll.metter.exception.PollNotFoundException;
 import com.ibm.ro.tm.apprenticeship.poll.metter.exception.UserNotFoundException;
 import com.ibm.ro.tm.apprenticeship.poll.metter.repository.AnswerRepository;
 import com.ibm.ro.tm.apprenticeship.poll.metter.repository.PollRepository;
 import com.ibm.ro.tm.apprenticeship.poll.metter.repository.UserRepository;
+
+import dto.AnswerDto;
 
 /**
  * @author vlads
@@ -43,24 +46,49 @@ public class AnswerService {
 //		return answerRepository;
 //	}
 //	
-	public Answer add(Answer answer) {
-		return answerRepository.save(answer);
+	public Answer add(AnswerDto answerDto) {
+		Answer newAnswer = null;
+		newAnswer = new Answer(answerDto.getContent());
+		User user = userRepository.findById(answerDto.getUserId()).orElseThrow(() -> new UserNotFoundException("user id not found "+answerDto.getUserId()));
+		 Poll poll = pollRepository.findById(answerDto.getPollId()).orElseThrow(() -> new PollNotFoundException("poll id not found "+answerDto.getPollId()));
+		newAnswer.setUser(user);
+		newAnswer.setPoll(poll);
+		newAnswer.setVottingDetails(answerDto.getVottingDetail());		
+		return newAnswer;
+		
 	}
 //	
 	public List<Answer> findAllAnswers(){
 		return answerRepository.findAll();
 	}
-//	
-//	public Answer findById(Long id) {
-//		return answerRepository.findById(id).orElseThrow(()-> new AnswerNotFoundException("Answer not found"));
-//	}
-//	
-	public Answer update(Answer answer) {
-		return answerRepository.save(answer);
+	
+	public Answer findById(Long id) {
+		return answerRepository.findById(id).orElseThrow(()-> new AnswerNotFoundException("Answer not found"));
+	}
+	
+	public Answer update(AnswerDto answerDto) {
+		
+		if(userRepository.findById(answerDto.getUserId()).isPresent() && pollRepository.findById(answerDto.getPollId()).isPresent()) {
+			Answer updateAnswer = null;
+			updateAnswer = new Answer(answerDto.getContent());
+			User user = userRepository.findById(answerDto.getUserId()).get();
+			Poll poll = pollRepository.findById(answerDto.getPollId()).get();
+			updateAnswer.setUser(user);
+			updateAnswer.setPoll(poll);
+			updateAnswer.setVottingDetails(answerDto.getVottingDetail());
+			return updateAnswer;			
+		} else {
+			throw new AnswerNotFoundException("No answer existent");
+		}	
+
 	}
 	
 	public void delete(Long id) {
+		if(answerRepository.findById(id).isPresent()) {
 		answerRepository.deleteById(id);
+		} else {
+			throw new AnswerNotFoundException("Answer id "+id+" is not found");
+		}
 	}
